sidebar counts fixed
